app/requirements.txt
fastapi==0.115.5
pydantic==2.9.2
python-dotenv==1.0.1
anthropic>=0.40.0
supabase>=2.9.0
openai>=1.54.0
mem0ai>=0.1.0
